add badge related stuff
-  automatically create gamification interface when a new end user is created
   this uses django signals
-  automatically create new badges for a new user
-  assign the new user badge to a new user
-  refactor management command
